测试版本

+++ app/src/main/java/com/jimei/xiaolumeimei/model/ProductListBean.java
@@ -99,7 +99,7 @@ public class ProductListBean {
     @SerializedName("head_img") private String mHeadImg;
     @SerializedName("is_saleopen") private boolean mIsSaleopen;
     @SerializedName("is_newgood") private boolean mIsNewgood;
-    @SerializedName("std_sale_price") private int mStdSalePrice;
+    @SerializedName("std_sale_price") private double mStdSalePrice;
     @SerializedName("agent_price") private double mAgentPrice;
     @SerializedName("sale_time") private String mSaleTime;
     @SerializedName("offshelf_time") private Object mOffshelfTime;
@@ -254,7 +254,7 @@ public class ProductListBean {
       return mIsNewgood;
     }
 
-    public int getStdSalePrice() {
+    public double getStdSalePrice() {
       return mStdSalePrice;
     }
